% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.mia}
\alias{print.mia}
\title{Print method for objects of class "mia"}
\usage{
\method{print}{mia}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "mia".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments (ignored).}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "mia"
}
\examples{
res <- mia_nice(data = dat.sim,
                X_names = c("X1", "X2"),
                X_values_1 = c(0, 1), X_values_2 = c(0, 0),
                Y_model = Y ~ W + X1 + X2, W_model = W ~ X1 + X2)
print(res)

}
\seealso{
\code{\link{mia_nice}}, \code{\link{mia_ice}},
\code{\link{mia_aipw}}
}
